refactored to reporters
