+++ test/unit/simple_cov_report_test.rb
@@ -1,6 +1,6 @@
 require File.expand_path('../test_helper', File.dirname(__FILE__))
 
-class ReporterTest < Test::Unit::TestCase
+class SimpleCovReportTest < Test::Unit::TestCase
 
   test "report data" do
     Coverband.configure do |config|
@@ -8,7 +8,7 @@ class ReporterTest < Test::Unit::TestCase
       config.reporter          = 'std_out'
     end
 
-    Coverband::Reporter.expects(:current_root).returns('/root_dir')
+    Coverband::Reporters::SimpleCovReport.expects(:current_root).returns('/root_dir')
     fake_redis.expects(:smembers).with('coverband').returns(fake_coverband_members)
     
     fake_coverband_members.each do |key|
@@ -22,7 +22,7 @@ class ReporterTest < Test::Unit::TestCase
     
     Coverband.configuration.logger.expects('info').with(all_of(*matchers))
 
-    Coverband::Reporter.report
+    Coverband::Reporters::SimpleCovReport.report
   end
 
     test "report data with scov" do
@@ -32,7 +32,7 @@ class ReporterTest < Test::Unit::TestCase
       config.s3_bucket         = nil
     end
 
-    Coverband::Reporter.expects(:current_root).at_least_once.returns('/root_dir')
+    Coverband::Reporters::SimpleCovReport.expects(:current_root).at_least_once.returns('/root_dir')
     fake_redis.expects(:smembers).with('coverband').returns(fake_coverband_members)
     
     fake_coverband_members.each do |key|
@@ -43,7 +43,7 @@ class ReporterTest < Test::Unit::TestCase
     
     Coverband.configuration.logger.stubs('info')
 
-    Coverband::Reporter.report(open_report: false)
+    Coverband::Reporters::SimpleCovReport.report(open_report: false)
   end
 
   ####
@@ -63,7 +63,7 @@ class ReporterTest < Test::Unit::TestCase
     #the code takes config.root expands and adds a '/' for the final path in roots
     roots = ["/app/", '/full/remote_app/path/']
 
-    assert_equal "/full/remote_app/path/is/a/path.rb", Coverband::Reporter.filename_from_key(key, roots)
+    assert_equal "/full/remote_app/path/is/a/path.rb", Coverband::Reporters::SimpleCovReport.filename_from_key(key, roots)
   end
 
   test "filename_from_key fix filename a changing deploy path with double quotes" do
@@ -78,7 +78,7 @@ class ReporterTest < Test::Unit::TestCase
     # note to get regex to work for changing deploy directories it must be double escaped in double quotes or use single qoutes
     roots = ["/box/apps/app_name/releases/\\d+/", '/full/remote_app/path/']
 
-    assert_equal "/full/remote_app/path/app/models/user.rb", Coverband::Reporter.filename_from_key(key, roots)
+    assert_equal "/full/remote_app/path/app/models/user.rb", Coverband::Reporters::SimpleCovReport.filename_from_key(key, roots)
   end
 
   test "filename_from_key fix filename a changing deploy path with single quotes" do
@@ -93,7 +93,7 @@ class ReporterTest < Test::Unit::TestCase
     # note to get regex to work for changing deploy directories it must be double escaped in double quotes or use single qoutes
     roots = ['/box/apps/app_name/releases/\d+/', '/full/remote_app/path/']
 
-    assert_equal "/full/remote_app/path/app/models/user.rb", Coverband::Reporter.filename_from_key(key, roots)
+    assert_equal "/full/remote_app/path/app/models/user.rb", Coverband::Reporters::SimpleCovReport.filename_from_key(key, roots)
   end
 
   test "filename_from_key leave filename from a key with a local path" do
@@ -107,7 +107,7 @@ class ReporterTest < Test::Unit::TestCase
     #the code takes config.root expands and adds a '/' for the final path in roots
     roots = ["/app/", '/full/remote_app/path/']
 
-    assert_equal "/full/remote_app/path/is/a/path.rb", Coverband::Reporter.filename_from_key(key, roots)
+    assert_equal "/full/remote_app/path/is/a/path.rb", Coverband::Reporters::SimpleCovReport.filename_from_key(key, roots)
   end
 
   test "line_hash gets correct hash entry for a line key" do
@@ -129,7 +129,7 @@ class ReporterTest < Test::Unit::TestCase
     File.stubs(:exists?).returns(true)
     File.stubs(:foreach).returns(['line 1','line2','line3','line4','line5','line6'])
     
-    assert_equal expected, Coverband::Reporter.line_hash(current_redis, key, roots)
+    assert_equal expected, Coverband::Reporters::SimpleCovReport.line_hash(current_redis, key, roots)
   end
 
   test "line_hash adjusts relative paths" do
@@ -151,7 +151,7 @@ class ReporterTest < Test::Unit::TestCase
     File.stubs(:exists?).returns(true)
     File.stubs(:foreach).returns(['line 1','line2','line3','line4','line5','line6'])
     
-    assert_equal expected, Coverband::Reporter.line_hash(current_redis, key, roots)
+    assert_equal expected, Coverband::Reporters::SimpleCovReport.line_hash(current_redis, key, roots)
   end
 
   test "#merge_arrays basic merge preserves order and counts" do
@@ -159,7 +159,7 @@ class ReporterTest < Test::Unit::TestCase
     second = [nil,0,1,0,0]
     expects = [0,0,1,0,1]
 
-    assert_equal expects, Coverband::Reporter.merge_arrays(first, second)
+    assert_equal expects, Coverband::Reporters::SimpleCovReport.merge_arrays(first, second)
   end
 
   test "#merge_arrays basic merge preserves order and counts different lenths" do
@@ -167,7 +167,7 @@ class ReporterTest < Test::Unit::TestCase
     second = [nil,0,1,0,0,0,0,1]
     expects = [0,0,1,0,1,0,0,1]
 
-    assert_equal expects, Coverband::Reporter.merge_arrays(first, second)
+    assert_equal expects, Coverband::Reporters::SimpleCovReport.merge_arrays(first, second)
   end
 
   private
